Rearrange  utils
