+++ arrakis/utils/exceptions.py
@@ -1,0 +1,24 @@
+class Error(OSError):
+    pass
+
+
+class SameFileError(Error):
+    """Raised when source and destination are the same file."""
+
+
+class SpecialFileError(OSError):
+    """Raised when trying to do a kind of operation (e.g. copying) which is
+    not supported on a special file (e.g. a named pipe)"""
+
+
+class ExecError(OSError):
+    """Raised when a command could not be executed"""
+
+
+class ReadError(OSError):
+    """Raised when an archive cannot be read"""
+
+
+class RegistryError(Exception):
+    """Raised when a registry operation with the archiving
+    and unpacking registeries fails"""
